requirements.txt
flask
python-dotenv
ibm-watson
openai
ibm-watsonx-ai
